aider: Replace incorrect module reference with variable reference for network_interface_ids in modules/vm/main.tf.
# Aider chat conversation:

USER: I ran this command:

 terraform plan

And got this output:

[31m╷[0m[0m
[31m│[0m [0m[1m[31mError: [0m[0m[1mReference to undeclared module[0m
[31m│[0m [0m
[31m│[0m [0m[0m  on modules/vm/main.tf line 8, in resource "azurerm_linux_virtual_machine" "pernik_vm":
[31m│[0m [0m   8:   network_interface_ids = [[4mmodule.network.network_interface_id[0m[count.index]][0m
[31m│[0m [0m
[31m│[0m [0mNo module call named "network" is declared in module.vm.
[31m╵[0m[0m

ASSISTANT: The error message indicates that within the `modules/vm/main.tf` file, there is a reference to a module named "network" which is not declared within the `module.vm` scope. To resolve this, we need to ensure that the `network_interface_ids` attribute is correctly referencing the output from the `network` module that is declared in the root `main.tf` file.

Here is the diff to correct the reference:

```diff
--- modules/vm/main.tf
+++ modules/vm/main.tf
@@ ... @@
 resource "azurerm_linux_virtual_machine" "pernik_vm" {
   count = var.vm_count
   name                  = "${var.resource_group_name}-vm${count.index}"
   resource_group_name   = var.resource_group_name
   location              = var.location
   size                  = var.vm_size
   admin_username        = var.admin_username
-  network_interface_ids = [module.network.network_interface_id[count.index]]
+  network_interface_ids = [var.network_interface_id[count.index]]
   disable_password_authentication = true
   admin_ssh_key {
     username   = var.admin_username
     public_key = azurerm_ssh_public_key.pernik_key.public_key
   }
   ...
 }
```

This change replaces the incorrect module reference with a variable reference, which should be provided by the root module when calling the `vm` module. The variable `network_interface_id` should be passed from the root module to the `vm` module as an input variable.