aider: Fix variable type and reference for network_interface_id in vm module.
# Aider chat conversation:

USER: I ran this command:

 terraform plan

And got this output:

[31m╷[0m[0m
[31m│[0m [0m[1m[31mError: [0m[0m[1mInvalid index[0m
[31m│[0m [0m
[31m│[0m [0m[0m  on modules/vm/main.tf line 8, in resource "azurerm_linux_virtual_machine" "pernik_vm":
[31m│[0m [0m   8:   network_interface_ids = [var.network_interface_id[4m[count.index][0m][0m
[31m│[0m [0m    [90m├────────────────[0m
[31m│[0m [0m[0m    [90m│[0m [1mcount.index[0m is a number
[31m│[0m [0m[0m    [90m│[0m [1mvar.network_interface_id[0m is a string
[31m│[0m [0m[0m
[31m│[0m [0mThis value does not have any indices.
[31m╵[0m[0m

ASSISTANT: The error you're encountering is due to an incorrect reference to `var.network_interface_id` in the `modules/vm/main.tf` file. The variable `network_interface_id` is expected to be a list, but it is currently defined as a string in `modules/vm/variables.tf`. To fix this, we need to update the variable definition in `modules/vm/variables.tf` to expect a list of strings instead of a single string. Additionally, we need to update the reference in `modules/vm/main.tf` to use the correct variable.

Here are the diffs for those changes:

```diff
--- modules/vm/variables.tf
+++ modules/vm/variables.tf
@@ ... @@
 variable "network_interface_id" {
-  description = "ID of the network interface"
-  type        = string
+  description = "IDs of the network interfaces"
+  type        = list(string)
 }
@@ ... @@
```

```diff
--- modules/vm/main.tf
+++ modules/vm/main.tf
@@ ... @@
 resource "azurerm_linux_virtual_machine" "pernik_vm" {
   count = var.vm_count
   name                  = "${var.resource_group_name}-vm${count.index}"
   resource_group_name   = var.resource_group_name
   location              = var.location
   size                  = var.vm_size
   admin_username        = var.admin_username
-  network_interface_ids = [var.network_interface_id[count.index]]
+  network_interface_ids = var.network_interface_id
   disable_password_authentication = true
   admin_ssh_key {
     username   = var.admin_username
     public_key = azurerm_ssh_public_key.pernik_key.public_key
   }
   ...
 }
@@ ... @@
```

Apply these changes to the respective files, and then run `terraform plan` again to see if the issue is resolved.