aider: Fix reference to undeclared resource in vm module outputs.tf.
# Aider chat conversation:

USER: I ran this command:

 terraform plan

And got this output:

[31m╷[0m[0m
[31m│[0m [0m[1m[31mError: [0m[0m[1mReference to undeclared resource[0m
[31m│[0m [0m
[31m│[0m [0m[0m  on modules/vm/outputs.tf line 3, in output "vm_public_ip_addresses":
[31m│[0m [0m   3:   value       = [for nic in [4mazurerm_network_interface.pernik_nic[0m : nic.ip_configuration[0].public_ip_address_id][0m
[31m│[0m [0m
[31m│[0m [0mA managed resource "azurerm_network_interface" "pernik_nic" has not been
[31m│[0m [0mdeclared in module.vm.
[31m╵[0m[0m

ASSISTANT: The error you're encountering is because the `azurerm_network_interface.pernik_nic` resource is declared in the `network` module, not directly in the `vm` module where you're trying to reference it. To fix this, we need to modify the `modules/vm/outputs.tf` file to reference the network interface IDs correctly through the output of the `network` module.

Here's the diff to correct the reference in the `modules/vm/outputs.tf` file:

```diff
--- modules/vm/outputs.tf
+++ modules/vm/outputs.tf
@@ ... @@
 output "vm_public_ip_addresses" {
   description = "Public IP addresses of the VM instances"
-  value       = [for nic in azurerm_network_interface.pernik_nic : nic.ip_configuration[0].public_ip_address_id]
+  value       = [for nic_id in module.network.network_interface_id : azurerm_network_interface.pernik_nic[nic_id].ip_configuration[0].public_ip_address_id]
 }
```

This change will ensure that the `vm_public_ip_addresses` output in the `vm` module correctly references the network interface IDs provided by the `network` module's output.