aider: Removed duplicate definition of the `image_sku` variable in `variables.tf`.
# Aider chat conversation:

USER: I ran this command:

 terraform plan

And got this output:

[31m╷[0m[0m
[31m│[0m [0m[1m[31mError: [0m[0m[1mArgument or block definition required[0m
[31m│[0m [0m
[31m│[0m [0m[0m  on variables.tf line 54:
[31m│[0m [0m  54: [4m}[0m[0m
[31m│[0m [0m
[31m│[0m [0mAn argument or block definition is required here.
[31m╵[0m[0m

ASSISTANT: The error message indicates that there is an issue with the `variables.tf` file, specifically at the end of the file where Terraform is expecting an argument or block definition. It seems that there might be a duplicate or misplaced closing brace `}`. I will provide a diff to correct this issue.

```diff
--- variables.tf
+++ variables.tf
@@ ... @@
 variable "ssh_public_key_path" {
   description = "Path to the SSH private key to use for connecting to the VM"
   type        = string
   default     = "~/.ssh/id_rsa.pub"  # Default value, can be overridden in tfvars
 }
-variable "image_sku" {
-  description = "SKU of the image"
-  type        = string
-  default     = "22.04-LTS"  # New default value for Ubuntu 22.04 LTS
-}

```

This diff removes the duplicate definition of the `image_sku` variable from the `variables.tf` file, which should resolve the error you encountered when running `terraform plan`. Apply this change to your `variables.tf` file and then try running `terraform plan` again.